feat(uptime): add request metrics for hyper/reqwest (#1)
This PR introduces the ConnectionStats and RequestStats structs (and all the associated promise-plumbing), as well as the request-level stats tracking.

(hyper-util and hyper-tls depend on hyper, which are responsible for the connection-level work. reqwest is the api-level bit, but also some of the glue that binds the connection-oriented work to the request-oriented work.)